requirements.txt
GitPython>=3.1.43
pathvalidate>=3.2.0
psutil>=5.9.8
python-dotenv>=1.0.1
requests>=2.31.0
ruamel.yaml>=0.18.6
tqdm>=4.66.2
setuptools>=69.5.1